requirements.txt
asgiref==3.3.4
Django
djangorestframework
pytz==2021.1
sqlparse==0.4.1
